+++ src/jobs/jobs.controller.spec.ts
@@ -121,6 +121,7 @@ describe("JobsController", () => {
               port: configService.get<string>("NEO4J_PORT"),
               scheme: configService.get<string>("NEO4J_SCHEME"),
               username: configService.get<string>("NEO4J_USERNAME"),
+              database: configService.get<string>("NEO4J_DATABASE"),
             } as Neo4jConnection),
         }),
       ],